Refactored project - renamed files and classes
