--- joystick.py
@@ -1,17 +1,0 @@
-#
-# Copyright 2019 Games Creators Club
-#
-# MIT License
-#
-
-
-class Joystick:
-
-    def __init__(self):
-        pass
-
-    def readAxis(self):
-        raise NotImplementedError()
-
-    def readButtons(self):
-        raise NotImplementedError()
