+++ src/python/README.md
@@ -1,0 +1,1 @@
+# Build and such
