+++ src/python/bt_joystick/__init__.py
@@ -1,0 +1,17 @@
+
+#
+# Copyright 2019 Games Creators Club
+#
+# MIT License
+#
+
+
+
+import bt_joystick.bt_device_classes
+import bt_joystick.bt_device
+import bt_joystick.hid_report_descriptor
+import bt_joystick.sdp_record
+
+from bt_joystick.bt_device import BTDevice
+from bt_joystick.main import Joystick, BluetoothJoystickDeviceMain
+
