+++ src/python/bt_joystick/bt_device.py
@@ -13,17 +13,15 @@ import sys
 import time
 import socket
 
-import sdp_record
-import hid_report_descriptor
-
 import dbus.service
 from dbus.mainloop.glib import DBusGMainLoop
 
-from joystick import Joystick
+from bt_joystick import sdp_record
+from bt_joystick import hid_report_descriptor
 
-from bt_device_classes import LIMITED_DISCOVERABLE_MODE, PERIPHERAL, GAMEPAD
-from hid_report_descriptor import Usage
-from sdp_record import MinorDeviceClass
+from bt_joystick.bt_device_classes import LIMITED_DISCOVERABLE_MODE, PERIPHERAL, GAMEPAD
+from bt_joystick.hid_report_descriptor import Usage
+from bt_joystick.sdp_record import MinorDeviceClass
 
 
 class BTDevice(dbus.service.Object):
@@ -135,85 +133,3 @@ class BTDevice(dbus.service.Object):
 
     def send_message(self, message):
         self.cinterrupt.send(message)
-
-
-if __name__ == "__main__":
-    if not os.geteuid() == 0:
-        sys.exit("Only root can run this script")
-
-    DBusGMainLoop(set_as_default=True)
-
-    bt = BTDevice()
-    joystick = Joystick()
-
-    while True:
-        re_start = False
-
-        print("Waiting for connections")
-        bt.listen()
-
-        button_bits_1 = 0
-        button_bits_2 = 0
-
-        axis = [0, 0, 0, 0]
-        new_axis = [0, 0, 0, 0]
-
-        while not re_start:
-            time.sleep(0.1)
-
-            joystick_axis = joystick.readAxis()
-            joystick_buttons = joystick.readButtons()
-
-            new_axis[0] = joystick_axis['x']
-            new_axis[1] = joystick_axis['y']
-            new_axis[2] = joystick_axis['rx']
-            new_axis[3] = joystick_axis['ry']
-
-            new_button_bits_1 = 0
-            new_button_bits_2 = 0
-
-            # 'trigger', 'tl', 'tr', 'thumb', 'dpad_up', 'dpad_down', 'dpad_left', 'dpad_right', 'thumbl', 'thumbr'
-            if joystick_buttons['dpad_up']:
-                new_button_bits_1 |= 1
-            if joystick_buttons['dpad_down']:
-                new_button_bits_1 |= 2
-            if joystick_buttons['dpad_left']:
-                new_button_bits_1 |= 4
-            if joystick_buttons['dpad_right']:
-                new_button_bits_1 |= 8
-
-            if joystick_buttons['trigger']:
-                new_button_bits_1 |= 16
-            if joystick_buttons['tl']:
-                new_button_bits_1 |= 32
-            if joystick_buttons['tr']:
-                new_button_bits_1 |= 64
-            if joystick_buttons['thumb']:
-                new_button_bits_1 |= 128
-
-            if joystick_buttons['thumbl']:
-                new_button_bits_2 |= 1
-            if joystick_buttons['thumbr']:
-                new_button_bits_2 |= 2
-
-            has_changes = False
-
-            for i in range(0, 4):
-                if axis[i] != new_axis[i]:
-                    axis[i] = new_axis[i]
-                    has_changes = True
-
-            if button_bits_1 != new_button_bits_1 or button_bits_2 != new_button_bits_2:
-                button_bits_1 = new_button_bits_1
-                button_bits_2 = new_button_bits_2
-                has_changes = True
-
-            if has_changes:
-                data = bytes((0xA1, 0x01, button_bits_1, button_bits_2, axis[0], axis[1], axis[2], axis[3]))
-
-                # print("Changing data " + str(["{:02x}".format(d) for d in data]))
-                try:
-                    bt.send_message(data)
-                except Exception as e:
-                    print("Failed to send data - disconnected " + str(e))
-                    re_start = True
